sprint15

+++ backend/controllers/cards.js
@@ -1,7 +1,7 @@
 const Card = require('../models/card');
 
-
 const cardNotFoundHandler = () => {
+  console.log('cardNotFoundHandler')
   const error = new Error('document not found');
   error.statusCode = 404;
   error.name = 'UserNotFoundError';
@@ -18,7 +18,6 @@ module.exports.getCards = (req, res, next) => {
 module.exports.createCard = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
-  console.log('createcard controller', req.user)
   Card.create({ name, link, owner })
     .then((card) => {
       res.status(201).send(card)
@@ -28,11 +27,17 @@ module.exports.createCard = (req, res, next) => {
 
 module.exports.deleteCard = (req, res, next) => {
   const { cardId } = req.params;
-
-  Card.findByIdAndDelete(cardId)
+  Card.findById(cardId)
     .orFail(() => cardNotFoundHandler())
-    .then((card) => res.status(201).send(card))
-    .catch(next);
+    .then((card) => {
+      if (card.owner.equals(req.user._id)) {
+        Card.findByIdAndDelete(cardId)
+          .orFail(() => cardNotFoundHandler())
+          .then(card => res.status(201).send(card))
+          .catch(next)
+      }
+    })
+    .catch(next)
 };
 
 module.exports.likeCard = (req, res, next) => {
